requirements.txt
alembic==1.0.1
amqp==2.3.2
beautifulsoup4==4.6.3
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
ItsDangerous==1.0.0
Jinja2==2.10
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.5
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.6
redis==2.10.6
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
unbabel-py==0.49
urllib3==1.24
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
